requirements.txt
# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12

# LLM Providers
google-genai>=1.0.0
openai>=1.50.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
tiktoken>=0.7.0

# Data Validation & Config
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Async I/O
aiofiles>=24.1.0
httpx>=0.27.0

# Web Scraping & Ingestion
feedparser>=6.0.0
duckduckgo-search>=6.0.0
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# JSON Repair
json-repair>=0.30.0

# Document Parsing (optional, for document-source pipeline)
pymupdf>=1.24.0
openpyxl>=3.1.0
python-pptx>=1.0.0

# Auth
PyJWT>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
